Switched to Drive folder with RTF'd docs
Lots of issues with word docs being able to be open in different editors. Since the
source documents had a few validation issues, we decided to try to get those original
docs to a state where the VSCode OOXML validator (https://marketplace.visualstudio.com/items?itemName=mikeebowen.ooxml-validator-vscode)
showed 0 errors on the docs. Eventually landed on the fact that if you open the source documents
in Desktop Microsoft Word, export the documents to rich text (RTF), then open the RTF documents
back up in word again and save them as DOCXs, the errors go away. So we did that in bulk for all
of the documents and uploaded them to a new folder.