Updated "we're slow" text in docx
we're slow == "It collates into one downloadable document, TCLP clauses published on our website as at 7 April 2022. Clauses published and updated since 7 April 2022 are not yet included in the Climate Clause Selector. We are working on adding them to the Climate Clause Selector. In the meantime"

More details in the previous commit.