Add the cache refresh interview
So normal people can go in and manually remove gdocs and airtable
caches. You shouldn't really have to remove the airtable cache, as
the whole thing refreshes every hour, but for gdocs you might want to
if you have to revert a document to an older (i.e. older gdocs modified
time) version of the document. We won't automatically use the currently
available one, just the last modified one.